src/backend/requirements.txt
flask==1.1.2
hdfs==2.6.0
tweepy==3.10.0
kafka-python==2.0.2
pyspark==3.1.2
python-dotenv==0.17.1
